add ability to register and navigation directly to the dashboard
